// src/router/router-event.enum.ts
export enum RouterEvent {
    ROUTE_MATCH = 'ROUTE_MATCH',
    NAVIGATION_START = 'NAVIGATION_START',
    ACTIVATION_START = 'ACTIVATION_START',
    ACTIVATION_FAILED = 'ACTIVATION_FAILED',
    ACTIVATION_END = 'ACTIVATION_END',
    RESOLVE_START = 'RESOLVE_START',
    RESOLVE_FAILED = 'RESOLVE_FAILED',
    RESOLVE_END = 'RESOLVE_END',
    NAVIGATION_END = 'NAVIGATION_END',
    NAVIGATION_FAILED = 'NAVIGATION_FAILED'
}